stormtrooper/requirements.txt
git+https://github.com/centre-for-humanities-computing/stormtrooper
torch
pyyaml
python-Levenshtein